refractored Accordion
