+++ MTM-Frontend/src/components/Form/Accordions/AccordionSummary.tsx
@@ -1,0 +1,17 @@
+import { styled } from "@mui/material/styles";
+import MuiAccordionSummary, {
+  AccordionSummaryProps,
+} from "@mui/material/AccordionSummary";
+
+interface AccordionSummaryStyledProps extends AccordionSummaryProps {
+  customStyles?: React.CSSProperties;
+}
+
+export const AccordionSummary = styled(
+  ({ customStyles, ...props }: AccordionSummaryStyledProps) => (
+    <MuiAccordionSummary {...props} />
+  ),
+)(({ theme, customStyles }) => ({
+  padding: 0,
+  ...customStyles,
+}));
